back/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
flake8
pylint
passlib 
bcrypt
sqlalchemy 
pymysql
jwt
python-jose
python-multipart
